Add packaging='pom' to example BOM
Currently, for any "BOM" dependency, Gradle will treat all <dependencyManagement>
elements as dependency constraints. Any POM with packaging == 'pom' is
considered a BOM for this purpose.

In the future, we will consider treating all <dependencyManagement> elements
as dependency constraints, regardless of the POM packaging element. However
this will be a more disruptive change, and we need to assess performance
implications.